Remove the unused EncodingDecoder
The project "encoding" appears to be unmaintained, and hasn't received
an update since 2017. Unfortunately the "encoding" artifacts on
crates.io do not contain the licenses for the project, which adds
difficulty in auditing dependencies.

Since "tiny-http" doesn't actually use the EncodingDecoder, this patch
removes it and it's dependency on "encoding" in order to avoid
complications resolving the "tiny-http" dependency story.

Instead, I would suggest investigating migrating over to
https://docs.rs/encoding_rs/0.8.13/encoding_rs/, which appears to be
actively maintained by Mozilla.

If this patch is accepted, would it be possible to get a release as
well?