Remove test for prompt handling of requests with `Expect: 100-continue`
On reflection, this its not really possible for this test to pass.
The server can't (or rather, shouldn't) read the request body until it
has sent a `100 continue` response.  But it cant send a continue
response until it has finished sending the _full_ response to the
previous request in the pipeline, because responses have to be sent in
the correct order.  This forced interleaving of reads and writes prevents
prompt request handling.