Test against MSRV 1.48 rather than 1.41.1
The OpenSSL crate has introduced a breaking change with 0.10.36 which
menas they now rely on const fn support not available in 1.41.1.

As only one OpenSSL version can be supported in a single Rust project,
and the change was released as a patch increment so Cargo will always
select this as the desired version, there's not much point in us testing
tiny-http on such an old compiler version anymore.