Correctly format HTTPDate using time-rs
Format strings for time-rs are an original and more complex syntax than
any of the strptime inspired examples that preceeded it.

This syntax must be parsed before it can be used in a formatter, so we
invoke a proc-macro to do that at compile time rather than every time
the formatter is required.

As time-rs requires an MSRV of 1.51, we must also increase our minimum
supported version.