Remove the 'static requirement on TestRequest::path
With 'static, one can only use static string slices, which makes it
impossible to test with urls generated at runtime, such as ones
generated by a fuzzer.

This change is backwards-compatible, because it only relaxes the
requirements on `with_path`, and although the type of the field changed,
that field is not exposed.

This doesn't incur an extra copy either. The copy was already there,
it only got moved out of `into` and into `with_path` now.